plugin.yml updated...
I'm not sure this is correct. I generated the file in the root, the plugin shell generator put it in root, but my NetBeans environment doesn't see/include/compile it unless it's in src/.

This may or may not work for other people. I may well have screwed something up, or missed a configuration setting. Argh.